grosse update

+++ utilitaires/gestion_db.py
@@ -174,7 +174,7 @@ def sauvegarde_score_ordi(jeu:str, ordi:str, valeur:float) -> None:
     #Déclaration des variables
     conn: sqlite3.Connection
     cursor: sqlite3.Cursor
-    ordis: list[tuple[int, str, int, int, int]]
+    ordis: list[tuple[int, str, int, float, int]]
 
     #Vérification de l'existence du jeu
     if jeu not in ["allumettes", "morpion", "devinettes"]:
@@ -209,7 +209,7 @@ def sauvegarde_score_ordi(jeu:str, ordi:str, valeur:float) -> None:
 
 
 #La fonction charger_score permet de charger les scores des joueurs dans la base de données, en fonction d'un jeu
-def charger_joueurs_jeu(jeu:str) -> list[tuple[int, str, int, int, int, int]]:
+def charger_joueurs_jeu(jeu:str) -> list[tuple[int, str, float, float, float, int]]:
     """
     Fonction pour charger les scores des joueurs d'un jeu en particulier
 
@@ -223,7 +223,7 @@ def charger_joueurs_jeu(jeu:str) -> list[tuple[int, str, int, int, int, int]]:
     #Déclaration des variables
     conn: sqlite3.Connection
     cursor: sqlite3.Cursor
-    scores: list[tuple[int, str, int, int, int, int]]
+    scores: list[tuple[int, str, float, float, float, int]]
 
     #Vérification de l'existence du jeu
     if jeu not in ["allumettes", "morpion", "devinettes"]:
@@ -242,7 +242,7 @@ def charger_joueurs_jeu(jeu:str) -> list[tuple[int, str, int, int, int, int]]:
 
 
 #La fonction charger_score permet de charger les scores des ordinateurs dans la base de données, en fonction d'un jeu
-def charger_ordi_jeu(jeu:str, ordre:bool=True) -> list[tuple[int, str, int, int, int]]:
+def charger_ordi_jeu(jeu:str, ordre:bool=True) -> list[tuple[int, str, int, float, int]]:
     """
     Fonction pour charger les scores des ordinateurs d'un jeu en particulier
 
@@ -256,7 +256,7 @@ def charger_ordi_jeu(jeu:str, ordre:bool=True) -> list[tuple[int, str, int, int,
     #Déclaration des variables
     conn: sqlite3.Connection
     cursor: sqlite3.Cursor
-    scores: list[tuple[int, str, int, int, int]]
+    scores: list[tuple[int, str, int, float, int]]
     requete: str
 
     #Vérification de l'existence du jeu
